+++ api-server/src/app.js
@@ -6,15 +6,12 @@ const cors = require('cors');
 const morgan = require('morgan');
 
 // Esoteric Resources
-const errorHandler = require( './middleware/error.js');
-const notFound = require( './middleware/404.js' );
+const errorHandler = require( './middleware/error');
+const notFound = require( './middleware/404' );
 
-// Models
-const Products = require('./models/products/products.js');
-const products = new Products();
-
-const Categories = require('./models/categories/categories.js');
-const categories = new Categories();
+// Routes
+const categoryRoutes = require('./routes/category-routes');
+const productRoutes = require('./routes/product-routes');
 
 // Prepare the express app
 const app = express();
@@ -27,109 +24,12 @@ app.use(express.json());
 app.use(express.urlencoded({extended:true}));
 
 // Routes
-app.get('/api/v1/categories', getCategories);
-app.post('/api/v1/categories', postCategories);
-app.get('/api/v1/categories/:id', getCategory);
-app.put('/api/v1/categories/:id', putCategories);
-app.delete('/api/v1/categories/:id', deleteCategories);
-
-app.get('/api/v1/products', getProducts);
-app.post('/api/v1/products', postProducts);
-app.get('/api/v1/products/:id', getProduct);
-app.put('/api/v1/products/:id', putProducts);
-app.delete('/api/v1/products/:id', deleteProducts);
+app.use('/api/v1/categories', categoryRoutes);
+app.use('/api/v1/products', productRoutes);
 
 // Catchalls
 app.use(notFound);
 app.use(errorHandler);
-
-// ROUTE HANDLER FUNCTIONS
-function getCategories(request,response,next) {
-  // expects an array of object to be returned from the model
-  categories.get()
-    .then( data => {
-      const output = {
-        count: data.length,
-        results: data,
-      };
-      response.status(200).json(output);
-    })
-    .catch( next );
-}
-
-function getCategory(request,response,next) {
-  // expects an array with the one matching record from the model
-  categories.get(request.params.id)
-    .then( result => response.status(200).json(result[0]) )
-    .catch( next );
-}
-
-function postCategories(request,response,next) {
-  // expects the record that was just added to the database
-  categories.create(request.body)
-    .then( result => response.status(201).json(result) )
-    .catch( next );
-}
-
-
-function putCategories(request,response,next) {
-  // expects the record that was just updated in the database
-  categories.update(request.params.id, request.body)
-    .then( result => response.status(200).json(result) )
-    .catch( next );
-}
-
-function deleteCategories(request,response,next) {
-  // Expects no return value (resource was deleted)
-  categories.delete(request.params.id)
-    .then( result => response.status(200).json(result) )
-    .catch( next );
-}
-
-
-function getProducts(request,response,next) {
-  // expects an array of objects back
-  products.get()
-    .then( data => {
-      const output = {
-        count: data.length,
-        results: data,
-      };
-      response.status(200).json(output);
-    })
-    .catch( next );
-}
-
-function getProduct(request,response,next) {
-  // expects an array with one object in it
-  products.get(request.params.id)
-    .then( result => response.status(200).json(result[0]) )
-    .catch( next );
-}
-
-function postProducts(request,response,next) {
-  // expects the record that was just added to the database
-  products.create(request.body)
-    .then( result => response.status(201).json(result) )
-    .catch( next );
-}
-
-
-function putProducts(request,response,next) {
-  // expects the record that was just updated in the database
-  products.update(request.params.id, request.body)
-    .then( result => response.status(200).json(result) )
-    .catch( next );
-}
-
-function deleteProducts(request,response,next) {
-  // Expects no return value (the resource should be gone)
-  products.delete(request.params.id)
-    .then( result => response.status(200).json(result) )
-    .catch( next );
-}
-
-
 
 module.exports = {
   server: app,
